backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic-settings==2.8.1
python-dotenv==1.0.1
psycopg[binary]==3.2.13
httpx==0.28.1
<<<<<<< HEAD
=======
openai==1.55.3
requests==2.32.3
>>>>>>> work
